tests/intel/xe_mmap: Fix truncation of visible VRAM size
xe_visible_vram_size() returns a 64-bit value. Storing it in 32-bit
truncated sizes above 4 GiB and led to false positives.

Fixes: c93b4e6ff2f3 ("tests/xe/mmap: sanity check small-bar")
Cc: Lukasz Laguna <lukasz.laguna@intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>

+++ tests/intel/xe_mmap.c
@@ -291,7 +291,7 @@ static void trap_sigbus(uint32_t *ptr)
 static void test_small_bar(int fd)
 {
 	size_t page_size = xe_get_default_alignment(fd);
-	uint32_t visible_size = xe_visible_vram_size(fd, 0);
+	uint64_t visible_size = xe_visible_vram_size(fd, 0);
 	uint32_t bo;
 	uint64_t mmo;
 	uint32_t *map;
